Fixed boulder path on large map

+++ src/main.cpp
@@ -479,7 +479,7 @@ long apply_wallhug_force_to_boulder(struct Thing *thing)
         angle = thing->move_angle_xy;
         if ( (angle) && ( (angle <= ANGLE_EAST) || (angle > ANGLE_WEST) ) )
         {
-          unsigned short y = thing->mappos.y.val;
+          int32_t y = thing->mappos.y.val;
           pos2.x.val = thing->mappos.x.val;
           pos2.z.val = 0;
           pos2.y.val = y - STL_PER_SLB * speed;
@@ -508,7 +508,7 @@ long apply_wallhug_force_to_boulder(struct Thing *thing)
         }
         else
         {
-          unsigned short x = thing->mappos.x.val;
+          int32_t x = thing->mappos.x.val;
           pos2.z.val = 0;
           pos2.y.val = thing->mappos.y.val;
           pos2.x.val = x - STL_PER_SLB * speed;
